+++ src/hooks/ready/registerGameSettings.js
@@ -638,6 +638,13 @@ export default function () {
     default: false,
   });
 
+  game.settings.register("ddb-importer", "character-update-policy-ignore-non-ddb-items", {
+    scope: "player",
+    config: false,
+    type: Boolean,
+    default: false,
+  });
+
   game.settings.register("ddb-importer", "character-update-policy-use-override", {
     scope: "player",
     config: false,